Fix ChannelState#clean() to handle received_at as Date type
Because the client sets received_at as a Date type (src/client.ts#L1168)
and ChannelState#clean() was expecting a string, it hasn't been working
at all. A better fix might be to restrict the Event typing to Date since
it is never a string after being handled by the client, but in lieu of
that we'll handle both types here and add a test.

Signed-off-by: Evan Owen <kainosnoema@gmail.com>