fix: [REACT-353] unify pinned_at & archived_at nullish values (#1516)
## Description of the changes, What, Why and How?

Event handler for `member.updated` receives a `member` object which
contains `archived_at` and `pinned_at` with `null` values if
unarchive/unpin actions are called - to make matters for our integrators
easier, the only nullish value for these would be `undefined`. I'll keep
the types unchanged (#1515) as the `ChannelMemberResponse` used in the
`Event` type is still correct.