fix: upsert attachmemts only when attachment exist and handle FileList for react native (#1539)
The goal of the PR is to upsert attachments only when they exist. This
is especially important because on React Native, we allow cancelling
uploading by calling removeAttachments, but when the upload still
happens, the upsert proceeds. We will only update you if the attachment
is present.

The other fix is that the `FileList` is not supported on RN. Thanks to
@MartinCupela for a prompt suggestion on this issue.