feat: introduce sync search (#1558)
The goal of the PR is to introduce sync search by extending the current
API. The previously `SearchSource` is now changed to `ISearchSource`,
and the `SearchSource` is the async variant of `ISearchSource`, and the
`SearchSourceSync` is the sync variant of the `ISearchSource`.

The previously named `BaseSearchSource` is still intact and handles the
async search, and the `BaseSearchSourceSync` is introduced to provide
sync search capabilities. Both extend the `BaseSearchSourceBase` that
handles the common utilities required for search.

This change was made to ensure that the classes that previously extended
the `BaseSearchSource` do not complain about the new abstract methods if
they were introduced, because you will never have the same search source
having async and sync variants in the same class. Moreover, this change
is done in a non-breaking way.

NOTE: The names used right now are a bit lame, and I am open to changing
them to a suitable convention. Any suggestions would be highly
appreciated.