feat: handle offline support for drafts (#1559)
## CLA

- [x] I have signed the [Stream
CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform)
(required).
- [x] Code changes are tested

## Description of the changes, What, Why and How?

The goal of the PR is to handle offline support for draft messages and
also allow composing a message when the composition is not empty when
the offline DB is enabled. Thanks to @isekovanic for the help with the
offline support explanation here.

## Changelog

-