feat: query channels request override (#1583)
## CLA

- [ ] I have signed the [Stream
CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform)
(required).
- [ ] Code changes are tested

## Description of the changes, What, Why and How?

This PR introduces a way for integrators to override the default
`ChannelManager`'s `queryChannels` (which is essentially
`client.queryChannels`.

This is particularly useful in certain niche' edge cases, where for
example if querying channels relies heavily on `{ cid: { $in: [...] }}`
filters (or something similar) it would be impossible to properly
paginate through everything as changing the `filters` on the fly causes
the entire state to be wiped as it's considered to be a brand new query.

It also comes with a set of rules:

- `client.queryChannels` must always be invoked by this override (the
reason being that `queryChannels` additionally does a bunch of state
updates under the hood, so it's mandatory that this is run)
- The `filters` stay intact during all of this (if they change, the
`ChannelManager` will consider this to be a brand new query and wipe
everything)
- The override still needs to respect `limit` and `offset`, so that
pagination works properly (this comes as a given since we pin specific
IDs, but worth mentioning still)

## Changelog

-