fix: set loading state to false when query channel errors out (#1585)
The PR focuses on the following:
1. Set the loading state to false if the query channel API errors out.
This was essential as currently the channel loads infinitely.
2. Manage the query channels error state for offline support enabled
properly.
3. When the sync API returns an response timeout error from the BE or
from Axios we should not clear/reset the DB. Doing it was resulting in
empty channel list even if the data should be there in the DB. This
happened because we just cleared the DB.

Closes #1584
https://github.com/GetStream/stream-chat-react-native/issues/3180