fix: search cursor pagination (#1611)
## CLA

- [ ] I have signed the [Stream
CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform)
(required).
- [ ] Code changes are tested

## Description of the changes, What, Why and How?

This bug was actually discovered on the feeds SDK, where we use the same
search system and were seeing weird behaviours when using cursor
pagination.

Specifically, when:

- The `pageSize` matches the size of the result
- Pagination has ended (`next` is not returned)

the search queries would still run and `hasNext` would never be updated.
This is because it would fail to consider it as `cursor` pagination (as
`next` would be neither `null` nor truthy) and it would fallback to
`offset` pagination, however since `pageSize` would match the number of
items returned it would consider `hasNext` to still be `true`. This
would then cause the subsequent queries to not know that they should not
run and they will run with the last known `next` cursor (since it also
never updated to the new `undefined` value), causing an endless loop of
repeating the last page over and over.

In the response from the backend, `next` is never returned as a falsy
value; but rather it's either there (and is a `string`) or it is not
returned at all. Since all search source queries return `next`, we can
safely check if it exists or not and use this as leverage to figure out
whether it's `cursor` or `offset` pagination.

## Changelog

-