fix: do not add error to client notifications for getDrafts (#1631)
Currently, if you listen to client notifications on your app and move
from channel screen to thread screen. This will trigger the notification
if a draft is not present for a parent message id as
`this.channel.getDrafts` would error out. Ideally, this should be
gracefully handled without showing a client-side notification. This has
been fixed in the PR.