feat: add ephemeralUpdateMessage server-side operation (#1662)
Extends the server-side client with `ephemeralUpdateMessage` operation.
When this API is used, the "updates" aren't stored in the database, but
they do generate the `message.updated` event.
This is useful for a flow where a certain message is frequently updated
(e.g, integration with LLM).

To preserve the updates, a `client.partialUpdateMessage` must be
performed, as otherwise there will be a data loss.